requirements.txt
# Requirements for the completed project

# Core LangChain dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0

# OpenAI API
openai>=1.0.0

# Vector database and embeddings
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# Web interface
streamlit>=1.28.0

# Environment and utilities
python-dotenv>=1.0.0
numpy>=1.21.0

# Optional: For token counting (used by LangChain)
tiktoken>=0.5.0
